rabbitmq_project/requirements.txt
pika==1.2.0
prometheus-client==0.16.0
pytest==7.4.0
pytest-mock==3.12.0
python-dotenv==0.21.0
requests==2.28.1
psutil==5.9.4
retry==0.9.2
schedule==1.1.0
```